Cast prices to unsigned instead of float for better MySQL compatibility

+++ classes/index/mysql/MySQL.php
@@ -28,7 +28,7 @@ class MySQL implements Index
      * @var array
      */
     public $columnCasts = [
-        'prices'      => 'float',
+        'prices'      => 'unsigned',
         'sort_orders' => 'unsigned',
     ];
